+++ solid/solid/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @einfach/solid
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.10
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.9
+
 ## 0.2.8
 
 ### Patch Changes